Remove duplicate definitions of reinterpret*
