+++ src/webgpu/util/conversion.ts
@@ -1094,91 +1094,6 @@ export const True = bool(true);
 /** A 'false' literal value */
 export const False = bool(false);
 
-// Encoding to u32s, instead of BigInt, for serialization
-export function reinterpretF64AsU32s(f64: number): [number, number] {
-  workingDataF64[0] = f64;
-  return [workingDataU32[0], workingDataU32[1]];
-}
-
-// De-encoding from u32s, instead of BigInt, for serialization
-export function reinterpretU32sAsF64(u32s: [number, number]): number {
-  workingDataU32[0] = u32s[0];
-  workingDataU32[1] = u32s[1];
-  return workingDataF64[0];
-}
-
-/**
- * @returns a number representing the u32 interpretation
- * of the bits of a number assumed to be an f32 value.
- */
-export function reinterpretF32AsU32(f32: number): number {
-  workingDataF32[0] = f32;
-  return workingDataU32[0];
-}
-
-/**
- * @returns a number representing the i32 interpretation
- * of the bits of a number assumed to be an f32 value.
- */
-export function reinterpretF32AsI32(f32: number): number {
-  workingDataF32[0] = f32;
-  return workingDataI32[0];
-}
-
-/**
- * @returns a number representing the f32 interpretation
- * of the bits of a number assumed to be an u32 value.
- */
-export function reinterpretU32AsF32(u32: number): number {
-  workingDataU32[0] = u32;
-  return workingDataF32[0];
-}
-
-/**
- * @returns a number representing the i32 interpretation
- * of the bits of a number assumed to be an u32 value.
- */
-export function reinterpretU32AsI32(u32: number): number {
-  workingDataU32[0] = u32;
-  return workingDataI32[0];
-}
-
-/**
- * @returns a number representing the u32 interpretation
- * of the bits of a number assumed to be an i32 value.
- */
-export function reinterpretI32AsU32(i32: number): number {
-  workingDataI32[0] = i32;
-  return workingDataU32[0];
-}
-
-/**
- * @returns a number representing the f32 interpretation
- * of the bits of a number assumed to be an i32 value.
- */
-export function reinterpretI32AsF32(i32: number): number {
-  workingDataI32[0] = i32;
-  return workingDataF32[0];
-}
-
-/**
- * @returns a number representing the u16 interpretation
- * of the bits of a number assumed to be an f16 value.
- */
-export function reinterpretF16AsU16(f16: number): number {
-  workingDataF16[0] = f16;
-  return workingDataU16[0];
-}
-
-/**
- * @returns a number representing the f16 interpretation
- * of the bits of a number assumed to be an u16 value.
- */
-export function reinterpretU16AsF16(u16: number): number {
-  workingDataU16[0] = u16;
-  return workingDataF16[0];
-}
-
 /**
  * Class that encapsulates a vector value.
  */
