+++ src/webgpu/util/math.ts
@@ -6,7 +6,13 @@ import {
 } from '../../external/petamoriken/float16/float16.js';
 
 import { kBit, kValue } from './constants.js';
-import { floatBitsToNumber, i32, kFloat16Format, kFloat32Format, u32 } from './conversion.js';
+import { i32, u32 } from './conversion.js';
+import {
+  reinterpretF64AsU64,
+  reinterpretU64AsF64,
+  reinterpretU32AsF32,
+  reinterpretU16AsF16,
+} from './reinterpret.js';
 
 /**
  * A multiple of 8 guaranteed to be way too large to allocate (just under 8 pebibytes).
@@ -2059,38 +2065,6 @@ export function gcd(a: number, b: number): number {
 /** @returns the Least Common Multiplier (LCM) of the inputs */
 export function lcm(a: number, b: number): number {
   return (a * b) / gcd(a, b);
-}
-
-/**
- * @returns the bit representation as a 64-integer, via interpreting the input
- * as a 64-bit float value
- */
-export function reinterpretF64AsU64(input: number): bigint {
-  return new BigUint64Array(new Float64Array([input]).buffer)[0];
-}
-
-/**
- * @returns a 64-bit float value via interpreting the input as the bit
- * representation as a 64-bit integer
- */
-export function reinterpretU64AsF64(input: bigint): number {
-  return new Float64Array(new BigUint64Array([input]).buffer)[0];
-}
-
-/**
- * @returns a 32-bit float value via interpreting the input as the bit
- * representation as a 32-bit integer
- */
-export function reinterpretU32AsF32(input: number): number {
-  return floatBitsToNumber(input, kFloat32Format);
-}
-
-/**
- * @returns a 16-bit float value via interpreting the input as the bit
- * representation as a 16-bit integer
- */
-export function reinterpretU16AsF16(hex: number): number {
-  return floatBitsToNumber(hex, kFloat16Format);
 }
 
 /** @returns the cross of an array with the intermediate result of cartesianProduct
